Refactor RuleConverterAppExamples class and update test

+++ src/converter/app.examples.ts
@@ -11,7 +11,8 @@ const defaultRootPath = path.resolve("src", "rules");
 
 /**
  * Represents an example usage of the RuleConverterApp class.
- * This class is used to convert rules from the asset/test/rules.tsv file.
+ * Default behavior is converting rules from the asset/test/rules.tsv file and save them to the src/rules folder.
+ * If you want to convert rules from a different source or save them to a different destination, you can use this class as a starting point.
  */
 class RuleConverterAppExamples {
 
@@ -19,14 +20,26 @@ class RuleConverterAppExamples {
 
     /**
      * Constructs a new instance of the AppExamples class.
-     * 
+     * During construction, the following actions are performed:
+     * - The log and error files are deleted.
+     * - The rule converter service is initialized.
+     * - The destination root path is created if it does not exist.
+     * Log and error files are located in the .work folder.
+     * Remember to set OPENAI_API_KEY environment variable before running this app.
      * @param allow - A function that determines whether a rule is allowed based on its ID. Defaults to always true.
      * @param rootPath - The destinaton root path where the converted rules will be saved. Defaults to src/rules.
+     * @param deleteLogFiles - Indicates whether the log and error files should be deleted. Defaults to true. 
+     * If you want to avoid to reconvert already converted rules, set this parameter to false.
+     * @see convertRules
+     * @see RuleDaoExamples
+     * @see OpenAIRuleConverter
      */
-    constructor(allow: (ruleId: string) => boolean = defaultAllow, rootPath: string = defaultRootPath) {
-        const ruleDao: RuleDaoExamples = new RuleDaoExamples({ allow: allow });
-        fs.rmSync(ruleDao.logPath, { force: true });
-        fs.rmSync(ruleDao.errPath, { force: true });
+    constructor({ allow = defaultAllow, rootPath = defaultRootPath, deleteLogFiles = true }: { allow?: (ruleId: string) => boolean, rootPath?: string, deleteLogFiles?: boolean } = {}) {
+        const ruleDao: RuleDaoExamples = new RuleDaoExamples({ allow });
+        if (deleteLogFiles) {
+            fs.rmSync(ruleDao.logPath, { force: true });
+            fs.rmSync(ruleDao.errPath, { force: true });
+        }
         const ruleConverterService: IRuleConverterService = new OpenAIRuleConverter();
         if (!fs.existsSync(rootPath)) {
             fs.mkdirSync(rootPath);
@@ -36,12 +49,15 @@ class RuleConverterAppExamples {
     }
 
     /**
-     * Converts rules from the asset/test/rules.tsv file.
+     * Converts rules from the asset/test/rules.tsv file
      */
-    convertRules(): Promise<void> {
+    async convertRules(): Promise<void> {
         return this.app.convertRules();
     }
 
 }
 
 export { RuleConverterAppExamples }
+
+const app = new RuleConverterAppExamples({ deleteLogFiles: false });
+await app.convertRules();
